Destroy AshTestHelper in AshTestBase::TearDown
The instance is created in AshTestHelper::SetUp(), so destroy it in
AshTestBase::TearDown.
This provides a way to run some code after AshTestHelper's destruction
in inheriting fixtures in latter CLs.

There's two edge cases we cannot handle easily.
Screen() instance is updated in ash::Shell()'s tear down.

(Testing)ProfileManager needs to be destroyed "during" TearDown
to follow the production behavior (specifically, after destroying
ash::Shell(), but before SessionManager/UserManager).
As a workaround, this CL adds a callback to AshTestBase invoked just
before destroying AshTestHelper.

BUG=278643115
TEST=Tryjob

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6224054
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1415300}

+++ ash/test/ash_test_helper.cc
@@ -165,10 +165,12 @@ AshTestHelper::~AshTestHelper() {
 
   SimpleGeolocationProvider::DestroyForTesting();
 
-  // Ensure the next test starts with a null display::Screen.  This must be done
-  // here instead of in TearDown() since some tests test access to the Screen
-  // after the shell shuts down (which they use TearDown() to trigger).
-  ScreenAsh::DeleteScreenForShutdown();
+  if (destroy_screen_) {
+    // Ensure the next test starts with a null display::Screen.  This must be
+    // done here instead of in TearDown() since some tests test access to the
+    // Screen after the shell shuts down (which they use TearDown() to trigger).
+    ScreenAsh::DeleteScreenForShutdown();
+  }
 
   // This should never have a meaningful effect, since either there is no
   // ViewsTestHelperAura instance or the instance is currently in its
@@ -285,6 +287,7 @@ void AshTestHelper::SetUp(InitParams init_params) {
   create_global_cras_audio_handler_ =
       init_params.create_global_cras_audio_handler;
   create_quick_pair_mediator_ = init_params.create_quick_pair_mediator;
+  destroy_screen_ = init_params.destroy_screen;
 
   if (create_global_cras_audio_handler_) {
     // Create `CrasAudioHandler` for testing since `g_browser_process` is not
